dom/tests/mochitest/webcomponents/test_custom_element_namespace.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>

<window title="XUL Custom Elements"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="runTest();">
  <title>Custom Elements in a XUL document</title>

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>

  <script type="application/javascript">
  <![CDATA[
    SimpleTest.waitForExplicitFinish();

    const HTML_NS = "http://www.w3.org/1999/xhtml";
    const XUL_NS = "http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul";

    class TestXULCustomElement extends XULElement {
      constructor() {
        super();
      }

      get connected() {
        return true;
      }
    }

    customElements.define("test-xul-element", TestXULCustomElement);

    class TestHTMLCustomElement extends HTMLElement {
      constructor() {
        super();
      }

      get connected() {
        return true;
      }
    }

    customElements.define("test-html-element", TestHTMLCustomElement);

    class TestHTMLBuiltInElement extends HTMLButtonElement {
      constructor() {
        super();
      }

      get connected() {
        return true;
      }
    }

    customElements.define("html-built-in", TestHTMLBuiltInElement, { extends: "button" });

    function checkElement(element, ns, connected, type) {
      is(element.namespaceURI, ns, `${type} should have the correct namespace`);
      if (connected) {
        ok(element.connected, `${type} should have applied the class`);
      } else {
        is(element.connected, undefined, `${type} should not have applied the class`);
      }
    }

    function runTest() {
      let element = new TestXULCustomElement();
      checkElement(element, XUL_NS, true, "instantiated XUL");

      element = document.getElementById("xul1");
      checkElement(element, XUL_NS, true, "parsed XUL as XUL");

      element = document.getElementById("xul2");
      checkElement(element, HTML_NS, false, "parsed XUL as HTML");

      element = document.createElement("test-xul-element");
      checkElement(element, XUL_NS, true, "document.createElement(XUL)");

      element = document.createXULElement("test-xul-element");
      checkElement(element, XUL_NS, true, "document.createXULElement(XUL)");

      element = document.createElementNS(XUL_NS, "test-xul-element");
      checkElement(element, XUL_NS, true, "document.createElementNS(XUL, XUL)");

      element = document.createElementNS(HTML_NS, "test-xul-element");
      checkElement(element, HTML_NS, false, "document.createElementNS(HTML, XUL)");

      element = new TestHTMLCustomElement();
      checkElement(element, HTML_NS, true, "instantiated HTML");

      element = new TestHTMLBuiltInElement();
      checkElement(element, HTML_NS, true, "instantiated HTML");

      element = document.getElementById("html1");
      checkElement(element, XUL_NS, false, "parsed HTML as XUL");

      element = document.getElementById("html2");
      checkElement(element, HTML_NS, true, "parsed HTML as HTML");

      element = document.getElementById("html3");
      checkElement(element, HTML_NS, true, "parsed HTML as HTML");

      element = document.getElementById("html4");
      checkElement(element, HTML_NS, true, "parsed HTML as HTML");

      element = document.createElement("test-html-element");
      checkElement(element, XUL_NS, false, "document.createElement(HTML)");

      element = document.createXULElement("test-html-element");
      checkElement(element, XUL_NS, false, "document.createXULElement(HTML)");

      element = document.createElementNS(XUL_NS, "test-html-element");
      checkElement(element, XUL_NS, false, "document.createElementNS(XUL, HTML)");

      element = document.createElementNS(HTML_NS, "test-html-element");
      checkElement(element, HTML_NS, true, "document.createElementNS(HTML, HTML)");

      SimpleTest.finish();
    }
  ]]>
  </script>

  <test-xul-element id="xul1"/>
  <test-html-element id="html1"/>
  <button xmlns="http://www.w3.org/1999/xhtml" is="html-built-in" id="html4">Button</button>

  <body xmlns="http://www.w3.org/1999/xhtml">
    <p id="display"></p>
    <div id="content" style="display: none">
      <test-xul-element id="xul2"/>
      <test-html-element id="html2"/>
      <button is="html-built-in" id="html3">Button</button>
    </div>
    <pre id="test"></pre>
  </body>
</window>
